Transitions have been improved, and we fixed some bugs along the way.
Transitions (and patterns, for that matter) now apply independently to
to multiple strips. They were previously the same exact instance. For
patterns that's not much of an issue, but for transitions it definitely
is. Now they both work fine, we're using suppliers in the constructors.
This is a breaking change, but a fairly minor one.

Also, patterns are now fully compatible across segments of different
lengths! They already mostly were, but again transitions were raising
problems. Problems that have been solved, that is.